Added all the Views and ViewControllers up to the game scene, but without any logic so far

+++ L2P-Quiz/BackgroundFetchManager.h
@@ -11,7 +11,7 @@
 
 @interface BackgroundFetchManager : NSObject
 
-@property (nonatomic) int biggestIdParsed;
+@property (nonatomic) NSInteger biggestIdParsed;
 
 + (id)sharedManager;
